rename command to generate
