+++ src/commands/generate.ts
